I dont need addDragonToList method anymore, because i add them into players list in my FactoryStore method buyDragon.
Now, because i was getting a ConcurrentModificationException when trying to remove a dragon from the list of players dragons while looping through, i had to create a method to find a sick dragon (dragon whos health is below 0 percent.). then, i created a method removeSickDragonIfFound which removes sick dragon from players list of dragons.
To avoid a bunch of if sets checking size of dragonsOwned.size in my Game class, i check it in Player class methods.